v2.2 - more token fixes
Whoops! Having token-card relationships clarified on both individual cards and on the tokens themselves was resulting in multiple instances of token creation options available on the right click menu. This should be fixed now. Furthermore, the "Tracker -> Token" change has been reverted on the Projected tracker has been reverted as the problem seems to have been related to the formatting in the token association text.